Update 2.b.domestic_economic.R
Data frame with full CPI/Inflation. All three data sets are merged into one data frame. The value of inflation is listed in month 1 for every year and it has 0.0 as the value for months 2-12. I could not figure out how to properly interpolate that between years so i did not merge it into base data at the end.